Bug in Checkers. Other patterns work perfectly fine. Just Checkers. Why?

+++ RayTracer/src/Checkers.cpp
@@ -12,14 +12,15 @@ Checkers::Checkers(const Color& a, const Color& b, Matrix trans) : Pattern(trans
     color_a = a;
     color_b = b;
 }
+
 Color Checkers::at(const Tuple& pt) const {
     if(pt.type() != TupType::POINT){
         throw std::invalid_argument("pt must be a point in Checkers");
     }
-    int x = static_cast<int>(floor(pt[0]));
-    int y = static_cast<int>(floor(pt[1]));
-    int z = static_cast<int>(floor(pt[2]));
-    if(((x+y+z)%2 ) == 0){
+    int x = floor(pt[0]);
+    int y = floor(pt[1]);
+    int z = floor(pt[2]);
+    if(((static_cast<int>(x+y+z))%2 ) == 0){
         return color_a;
     }
     return color_b;
